requirements.txt
flask==2.3.2
numpy==1.24.3
scikit-learn==1.3.2
tensorflow-macos==2.13.0
tensorflow-metal==1.0.0
opencv-python==4.8.1.78
Pillow==10.0.0
joblib==1.3.2
tqdm==4.66.1
